Move trait classes to 'traits'
allocator_traits and block_allocator_traits were moved to a
new subdirectory 'traits' to remove a cyclic dependency.

Prior to this, utilities and concepts had a cyclic dependency wherein
concepts were made out of utility types, and [block_]allocator_traits
had a dependency to concepts. With this move, there is no more
intersubsystem cyclic dependency.
