+++ include/bit/memory/allocators/fallback_allocator.hpp
@@ -12,11 +12,12 @@
 #include "detail/named_allocator.hpp" // detail::named_allocator
 
 #include "../utilities/detail/ebo_storage.hpp" // detail::ebo_storage
-#include "../utilities/allocator_traits.hpp"   // allocator_traits
 #include "../utilities/macros.hpp"             // BIT_MEMORY_UNUSED
 #include "../utilities/owner.hpp"              // owner
 
 #include "../concepts/Allocator.hpp" // allocator_pointer_t, etc
+
+#include "../traits/allocator_traits.hpp" // allocator_traits
 
 #include <tuple>     // std::forward_as_tuple
 #include <utility>   // std::forward
